backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.16
django-cors-headers==4.4.0
django-jazzmin==3.0.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
frozenlist==1.5.0
idna==3.10
kombu==5.4.2
multidict==6.1.0
prompt-toolkit==3.0.48
propcache==0.2.0
PyJWT==2.9.0
reportlab==4.2.5
openpyxl==3.1.5
Pillow>=10.0.0
python-dateutil==2.9.0.post0
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
twilio==9.3.6
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.15.2
